Moved media upload settings into separate view
- New view for media upload settings
- Split settings into sections: Export, Auto Upload
- Changed appearance of the Upload Path setting
- Code refactoring changes
