+++ ownCloud/Settings/MediaUploadSettingsViewController.swift
@@ -1,0 +1,33 @@
+//
+//  MediaUploadSettingsViewController.swift
+//  ownCloud
+//
+//  Created by Michael Neuwert on 22.05.20.
+//  Copyright © 2020 ownCloud GmbH. All rights reserved.
+//
+
+/*
+ * Copyright (C) 2020, ownCloud GmbH.
+ *
+ * This code is covered by the GNU Public License Version 3.
+ *
+ * For distribution utilizing Apple mechanisms please see https://owncloud.org/contribute/iOS-license-exception/
+ * You should have received a copy of this license along with this program. If not, see <http://www.gnu.org/licenses/gpl-3.0.en.html>.
+ *
+ */
+
+import UIKit
+import ownCloudSDK
+
+class MediaUploadSettingsViewController: StaticTableViewController {
+
+	override func viewDidLoad() {
+		super.viewDidLoad()
+		navigationItem.title = "Media Upload".localized
+
+		if let userDefaults = OCAppIdentity.shared.userDefaults {
+			self.addSection(MediaExportSettingsSection(userDefaults: userDefaults))
+			self.addSection(AutoUploadSettingsSection(userDefaults: userDefaults))
+		}
+	}
+}
